+++ src/main/kotlin/com/github/alaanor/candid/inspection/CandidUnusedImportInspection.kt
@@ -1,21 +1,27 @@
 package com.github.alaanor.candid.inspection
 
-import com.github.alaanor.candid.quickfix.CandidRemoveUnusedImportQuickFix
+import com.github.alaanor.candid.quickfix.CandidRemoveImportQuickFix
 import com.github.alaanor.candid.util.CandidImportUtil
+import com.github.alaanor.candid.util.projectFilePath
 import com.intellij.codeInspection.*
 import com.intellij.psi.PsiFile
 
-class CandidUnusedImportInspection: LocalInspectionTool() {
+class CandidUnusedImportInspection : LocalInspectionTool() {
     override fun checkFile(file: PsiFile, manager: InspectionManager, isOnTheFly: Boolean): Array<ProblemDescriptor>? {
         val problemsHolder = ProblemsHolder(manager, file, isOnTheFly)
         val unusedImports = CandidImportUtil.getAllUnusedImportFor(file)
+        val currentFilePath = file.projectFilePath()
 
         unusedImports.forEach { unusedImport ->
+            if (unusedImport.projectFilePath() == currentFilePath) {
+                return@forEach
+            }
+
             problemsHolder.registerProblem(
                 unusedImport.originalElement,
                 "Unused import",
                 ProblemHighlightType.LIKE_UNUSED_SYMBOL,
-                CandidRemoveUnusedImportQuickFix(unusedImport)
+                CandidRemoveImportQuickFix(unusedImport, CandidRemoveImportQuickFix.Type.Unused)
             )
         }