+++ src/main/kotlin/com/github/alaanor/candid/quickfix/CandidRemoveImportQuickFix.kt
@@ -7,9 +7,20 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 
-class CandidRemoveUnusedImportQuickFix(import: CandidImportStatement) : LocalQuickFixOnPsiElement(import) {
-    override fun getText(): String = "Remove unused import"
-    override fun getFamilyName(): String = text
+class CandidRemoveImportQuickFix(import: CandidImportStatement, val type: Type) : LocalQuickFixOnPsiElement(import) {
+
+    enum class Type {
+        Invalid,
+        Unused,
+    }
+
+    override fun getText(): String {
+        return when (type) {
+            Type.Invalid -> "Remove invalid import"
+            Type.Unused -> "Remove unused import"
+        }
+    }
+    override fun getFamilyName(): String = "Remove import"
 
     override fun invoke(project: Project, file: PsiFile, startElement: PsiElement, endElement: PsiElement) {
         val element = startElement as? CandidImportStatement ?: return